% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdbscan.R
\name{st_dbscan}
\alias{st_dbscan}
\title{Spatio-Temporal DBSCAN}
\usage{
st_dbscan(data, eps_spatial, eps_temporal, min_pts, ...)
}
\arguments{
\item{data}{matrix. A matrix containing, \strong{in that order}, \code{x}, \code{y} and
\code{t}. \code{x} (longitude) and \code{y} (latitude) are the spatial coordinates and \code{t}
is the cumulative time since a common origin (\emph{e.g.} \code{c(0, 6, 10)}). \code{t}
must be sorted.}

\item{eps_spatial}{Numeric. The spatial radius threshold. Points closer than
this in space may belong to the same cluster.}

\item{eps_temporal}{Numeric. The temporal threshold. Points closer than this
in time may belong to the same cluster.}

\item{min_pts}{Integer. Minimum number of points required to form a core
point.}

\item{...}{Additional arguments are passed on to \code{dbscan::frNN()} and
\code{dbscan::dbscan()}.}
}
\value{
\code{st_dbscan()} returns an object of class \code{stdbscan} with the following
components:

\item{cluster }{Integer vector with cluster assignments. Zero indicates
noise points.}
\item{eps }{Value of the \code{eps_spatial} parameter.}
\item{minPts }{Value of the \code{minPts} parameter.}
\item{metric }{Used distance metric.}
\item{borderPoints }{Whether border points are considered as noise (\code{FALSE}) or not (\code{TRUE}).}
\item{eps_temporal }{Value of the \code{eps_temporal} parameter.}

This class is a simple extension of the \code{dbscan} class. For more details,
see \code{\link[dbscan:dbscan]{dbscan}} documentation.
}
\description{
Perform \strong{ST-DBSCAN} clustering on points with spatial and temporal
coordinates. This algorithm identifies clusters of points that are close both
in space and time.
}
\details{
ST-DBSCAN extends classical DBSCAN by incorporating a temporal constraint.
Two points are considered neighbors if they are within \code{eps_spatial} in
space \strong{and} within \code{eps_temporal} in time. Clusters are expanded from core
points recursively following the DBSCAN algorithm.

ST-DBSCAN is implemented using the following approach:

\enumerate{
\item Find the spatial neighbors using Fixed Radius Nearest Neighbors
(\code{dbscan::frNN()})
\item Filter the spatial neighbors by the temporal constraint
\item Apply DBSCAN on the filtered neighbors using \code{dbscan::dbscan()}
}
}
\examples{
data(geolife_traj)

geolife_traj$date_time <- as.POSIXct(
  paste(geolife_traj$date, geolife_traj$time),
  format = "\%Y-\%m-\%d \%H:\%M:\%S",
  tz = "GMT"
)
geolife_traj$t <- as.numeric(
  geolife_traj$date_time - min(geolife_traj$date_time)
)
data <- cbind(geolife_traj$x, geolife_traj$y, geolife_traj$t)

st_dbscan(
  data = data,
  eps_spatial = 3,
  eps_temporal = 30,
  min_pts = 3,
  # Extra arguments
  splitRule = "STD",
  search = "kdtree",
  approx = 1
)

}
\references{
Birant, D., & Kut, A. (2007). ST-DBSCAN: An algorithm for clustering
spatial–temporal data. Data & Knowledge Engineering, 60(1), 208–221.
https://doi.org/10.1016/j.datak.2006.01.013
}
