% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdbscan.R
\name{st_dbscan_corepoint}
\alias{st_dbscan_corepoint}
\title{Check if points are core points}
\usage{
st_dbscan_corepoint(data, eps_spatial, eps_temporal, min_pts, ...)
}
\arguments{
\item{data}{matrix. A matrix containing, \strong{in that order}, \code{x}, \code{y} and
\code{t}. \code{x} (longitude) and \code{y} (latitude) are the spatial coordinates and \code{t}
is the cumulative time since a common origin (\emph{e.g.} \code{c(0, 6, 10)}).}

\item{eps_spatial}{Numeric. The spatial radius threshold. Points closer than
this in space may belong to the same cluster.}

\item{eps_temporal}{Numeric. The temporal threshold. Points closer than this
in time may belong to the same cluster.}

\item{min_pts}{Integer. Minimum number of points required to form a core
point.}

\item{...}{Additional arguments are passed on to \code{dbscan::frNN()}.}
}
\value{
A boolean vector indicating if data points are core points.
}
\description{
Check if data points are core points. A core point is a point with more than
\code{min_pts} points in his neighborhood.
}
\examples{
data(geolife_traj)

geolife_traj$date_time <- as.POSIXct(
  paste(geolife_traj$date, geolife_traj$time),
  format = "\%Y-\%m-\%d \%H:\%M:\%S",
  tz = "GMT"
)
geolife_traj$t <- as.numeric(
  geolife_traj$date_time - min(geolife_traj$date_time)
)
data <- cbind(geolife_traj$x, geolife_traj$y, geolife_traj$t)

res <- st_dbscan_corepoint(
  data = data,
  eps_spatial = 3,
  eps_temporal = 30,
  min_pts = 3
)
head(res)

}
