% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdbscan-package.R
\docType{package}
\name{stdbscan-package}
\alias{stdbscan}
\alias{stdbscan-package}
\title{stdbscan: Spatio-Temporal DBSCAN Clustering}
\description{
Implements the ST-DBSCAN clustering algorithm for detecting spatially and temporally dense regions in point data, with a fast C++ backend via 'Rcpp'. Birant and Kut (2007) \doi{10.1016/j.datak.2006.01.013}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MiboraMinima/stdbscan/}
  \item \url{https://miboraminima.github.io/stdbscan/}
  \item Report bugs at \url{https://github.com/MiboraMinima/stdbscan/issues/}
}

}
\author{
\strong{Maintainer}: Antoine Le Doeuff \email{antoine.ldoeuff@gmail.com} (\href{https://orcid.org/0009-0008-8807-3816}{ORCID})

}
\keyword{internal}
