% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geolife_traj}
\alias{geolife_traj}
\title{GPS pings from the GeoLife GPS Trajectories dataset}
\format{
A data.frame with one row per ping and the following columns:

\describe{
\item{date}{(chr) the date
\item{time}{(chr) the time
\item{x}{(dbl) Longitude (EPSG:4586)
\item{y}{(dbl) Latitude (EPSG:4586)
}
}
\source{
https://www.microsoft.com/en-us/download/details.aspx?id=52367
}
\usage{
geolife_traj
}
\description{
Extraction of the GeoLife GPS Trajectories dataset. The selected trajectory
is 000-20081023025304.

Data manipulation applied to the raw data :
\itemize{
\item Convertion to EPSG:4586
\item Hand selection of pings
\item Selection of relevant variable
}
}
\keyword{datasets}
